Fixed bug that showed up when processing jQuery: I was using the "token in KEYWORDS" construct everywhere, and it doesn't work well because jQuery defines a lot of variables using names that have some standard meaning in JS, such as "toString".  "toString" in OBJECT will always evaluate to true, since all objects have a "toString" method, but it's not what we expect.
Take two was to replace all these constructs with foo.hasOwnProperty(bar),
which didn't work either for jQuery because it defines a variable called
"hasOwnProperty".  When parsing the scope, I put all the variable names in a
hash and later I'm calling that hash.hasOwnProperty(varname) to determine
wether the variable is in scope or not.  But "hasOwnProperty" was already
something else in that hash. :-)

Cool, isn't it.  So I added the HOP function which does:

    return Object.prototype.hasOwnProperty.call(hash, key);

and now we're flawless.