better fix for issue #128:
    (function(){
        var a = function fact(n){
            if (n == 1) return 1;
            else return n * fact(n - 1);
        };
        // console.log(fact(10)); // this errors out, `fact' not defined
        fact = 5;
        console.log(a(10)); // this works
    })();
    console.log(fact);

The above shows that when a function expression has a name, that name is not
introduced into the existing scope; instead it's only available in that
function's body (`fact' can call itself, but we can't call it from outside
by that name).  It also doesn't introduce a name, which is why setting fact
= 5 actually creates a global.

So UglifyJS is now able to mangle `fact' in the case above (or drop the name
in cases where the inner body is not referencing it).  The above mangles to:

    (function() {
        var a = function a(b) {
            return b == 1 ? 1 : b * a(b - 1);
        };
        fact = 5, console.log(a(10));
    })(), console.log(fact);