Fix broken calls of complex binary expressions [#471]
When deciding whether to parenthesize the expression being called, a
simple test of f.charAt(0) results in a broken code when calling a
complex expression which just happens to start with a paren but is
not actually parenthesized.

As an example, the code in issue471.js (see the diff) used to
incorrectly output `(a||b)&&c()`.